requirements.txt
streamlit==1.27.0
pillow